redesign architecture

+++ ykm_app/app/app.cpp
@@ -43,7 +43,7 @@ YkmApp_Result YkmApp_SetName(const char* name)
 
 YKM_APP_CAPI
 YkmApp_Result YkmApp_SetVersion( //
-    uint16_t big, uint16_t ver, uint32_t patch)
+    int big, int ver, int patch)
 {
     static std::string versionStr;
     if (versionStr.size() > 0)
